+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- added service report-license operation
+- 
+### Added
 - added custom_plugin_dir to .griffonrc which defines a custom directory for user plugins
 
 ## [0.1.6] - 2023-03-23